Reintegrate LCC demo save/load in the new qt5 version ok.
