+++ Linear_cell_complex/demo/Linear_cell_complex/typedefs.h
@@ -26,6 +26,12 @@
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Combinatorial_map_save_load.h>
 
+#include <CGAL/Triangulation_2_filtered_projection_traits_3.h>
+#include <CGAL/Triangulation_vertex_base_with_info_2.h>
+#include <CGAL/Triangulation_face_base_with_info_2.h>
+#include <CGAL/Constrained_Delaunay_triangulation_2.h>
+#include <CGAL/Constrained_triangulation_plus_2.h>
+
 #include <CGAL/IO/Color.h>
 #include <CGAL/Timer.h>
 #include <CGAL/Random.h>
@@ -168,6 +174,22 @@ typedef LCC::Point    Point_3;
 typedef LCC::Vector   Vector_3;
 
 typedef CGAL::Timer Timer;
+
+typedef CGAL::Triangulation_2_filtered_projection_traits_3<Mytraits> P_traits;
+typedef CGAL::Triangulation_vertex_base_with_info_2<Dart_handle,P_traits> Vb;
+
+struct Face_info {
+  bool exist_edge[3];
+  bool is_external;
+};
+
+typedef CGAL::Triangulation_face_base_with_info_2<Face_info,P_traits> Fb1;
+
+typedef CGAL::Constrained_triangulation_face_base_2<P_traits, Fb1>    Fb;
+typedef CGAL::Triangulation_data_structure_2<Vb,Fb>                   TDS;
+typedef CGAL::No_intersection_tag                                     Itag;
+typedef CGAL::Constrained_Delaunay_triangulation_2<P_traits, TDS,
+                                                   Itag>              CDT;
 
 struct Scene {
   LCC* lcc;
